Шаблонизация с учетом корректировок
