+++ data.php
@@ -1,0 +1,43 @@
+<?php
+
+$project_list = ['Входящие', 'Учеба', 'Работа', 'Домашние дела', 'Авто'];
+$task_list = [
+	[
+		'subject' => 'Собеседование в IT компании',
+		'date_complite' => '01.12.2018',
+		'catrgory' => 'Работа',
+		'performed' => 'Нет',
+	],	
+	[
+		'subject' => 'Выполнить тестовое задание',
+		'date_complite' => '25.12.2018',
+		'catrgory' => 'Работа',
+		'performed' => 'Нет',
+	],
+	[
+		'subject' => 'Сделать задание первого раздела',
+		'date_complite' => '21.12.2018',
+		'catrgory' => 'Учеба',
+		'performed' => 'Да',
+	],
+	[
+		'subject' => 'Встреча с другом',
+		'date_complite' => '22.12.2018',
+		'catrgory' => 'Входящие',
+		'performed' => 'Нет',
+	],
+	[
+		'subject' => 'Купить корм для кота',
+		'date_complite' => 'Нет',
+		'catrgory' => 'Домашние дела',
+		'performed' => 'Нет',
+	],
+	[
+		'subject' => 'Заказать пиццу',
+		'date_complite' => 'Нет',
+		'catrgory' => 'Домашние дела',
+		'performed' => 'Нет',
+	]
+];
+
+?>
